refactor: update import paths for consistency across modules
