+++ src/os_assistant/parsers/setup.py
@@ -4,7 +4,7 @@ from langchain.output_parsers import OutputFixingParser, PydanticOutputParser
 from langchain_core.exceptions import OutputParserException
 from langchain_core.prompts import PromptTemplate
 
-from src.os_assistant.utils.settings import fixing_model
+from os_assistant.utils.model_factory import fixing_model
 from os_assistant.parsers.json_cleaner import clean_and_parse_json
 from os_assistant.pydantic_models.schemas import (
     CodeExecuteRequest,
